requirements.txt
inference-sdk>=0.15.0
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
streamlit>=1.33.0
streamlit-image-coordinates>=0.1.9
plotly>=5.18.0
